requirements.txt
colorgram.py==1.2.0
Pillow==8.3.1
